requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.2
cryptography==2.1.4
Django==1.10.6
django-environ==0.4.4
Fabric3==1.13.1.post1
idna==2.6
Markdown==2.6.10
paramiko==2.4.0
pkg-resources==0.0.0
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pytz==2017.3
six==1.11.0
